Revert "ASPACE-291 WIP update tests"
This reverts commit 7d73484f93dfd46f4f8f066474241f1aa40f2208.

+++ lib/folio_sync/archives_space_to_folio/folio_synchronizer.rb
@@ -15,7 +15,6 @@ module FolioSync
         @downloads_location = downloads_location
       end
 
-      # last_x_hours is an optional parameter to limit the sync to resources modified in the last x hours
       def fetch_and_sync_resources_to_folio(last_x_hours)
         @downloading_errors = []
         @syncing_errors = []
@@ -68,9 +67,16 @@ module FolioSync
 
       # Location can be 'daily_sync' or 'manual_sync'
       def clear_downloads(location)
+        # puts "Clearing downloads directory for location: #{location}"
         config = Rails.configuration.folio_sync[:aspace_to_folio]
         downloads_dir = File.join(config[:marc_download_base_directory], @instance_key, location)
-        FileUtils.rm_rf(Dir["#{downloads_dir}/*"])
+
+        if Dir.exist?(downloads_dir)
+          FileUtils.rm_rf(Dir["#{downloads_dir}/*"])
+          puts "Cleared downloads directory: #{downloads_dir}"
+        else
+          puts "Downloads directory does not exist: #{downloads_dir}"
+        end
       end
     end
   end
